Hermitian Eigensystems

PRIMME.eigs — Function

PRIMME.eigs(A, k::Integer; kwargs...) => Λ,V,resids,stats

computes some (k) eigen-pairs of the n×n Hermitian matrix A

Returns Λ, a vector of k eigenvalues; V, a n×k matrix of right eigenvectors; resids, a vector of residual norms; and stats, a structure with an account of the work done.

Keyword args

(names in square brackes are analogues in PRIMME documentation)

which::Symbol, indicating which pairs should be computed. If a target sigma is provided, :SM and :LM refer to distance from the target, otherwise 'sigma' is implicitly zero.

:SR: smallest algebraic

:SM: smallest magnitude (default if sigma is provided)

:LR: largest algebraic

:LM: largest magnitude (default)

:CGT: closest greater than or equal to target

:CLT: closest less than or equal to target

sigma::Real, a target value; see which above.

tol: convergence criterion (residual norm relative to estimated norm of A) [eps]

maxiter: bound on outer iterations [maxOuterIterations]

maxMatvecs: bound on calls to matrix-vector multiplication function

P: preconditioning matrix or (preferably) factorization

B: mass matrix

verbosity::Int: detail of diagnostics to report to standard output [reportLevel]

check::Symbol: (:throw,:warn,:quiet) what to do on convergence failure

Some other keyword arguments, such as those passed to internal functions, might be properly documented someday.